Move get_commit_hunks to utils
This method is only used in packit (in update_source_git API method)
hence it can be safely moved. Since it doesn't combine work with a git
repo and its remote API but instead works with just the repo, a separate
helper function is a better place for this functionality.
