+++ packit/local_project.py
@@ -2,10 +2,9 @@
 # SPDX-License-Identifier: MIT
 
 import logging
-import re
 import shutil
 from pathlib import Path
-from typing import Optional, Union, Iterable, Iterator, List
+from typing import Optional, Union, Iterable, Iterator
 
 import git
 from git.exc import GitCommandError
@@ -570,25 +569,6 @@ class LocalProject:
 
     def get_commits(self, ref: str = "HEAD", **kwargs) -> Iterator[git.Commit]:
         return self.git_repo.iter_commits(ref, **kwargs)
-
-    def get_commit_hunks(self, commit: git.Commit) -> List[str]:
-        """Get a list of hunks of the given commit."""
-        patch = self.git_repo.git.show(commit, format="", color="never")
-        hunk_start_re = re.compile(r"diff --git a/.+ b/.+")
-        section_start = 0
-        result = []
-        patch_lines = patch.splitlines()
-        for i, line in enumerate(patch_lines):
-            if hunk_start_re.match(line):
-                section = patch_lines[section_start:i]
-                if section:
-                    result.append("\n".join(section))
-                section_start = i
-        # The last section
-        section = patch_lines[section_start:]
-        if section:
-            result.append("\n".join(section))
-        return result
 
     def fetch(self, remote: str, refspec: Optional[str] = None):
         """
