+++ packit/utils/repo.py
@@ -383,3 +383,32 @@ def get_commit_diff(commit: git.Commit) -> List[git.Diff]:
     else:
         # Probably a merge commit, we can't do much about it
         return []
+
+
+def get_commit_hunks(repo: git.Repo, commit: git.Commit) -> List[str]:
+    """Get a list of hunks of the given commit.
+
+    Args:
+        repo: Git repo which the commit belongs to.
+        commit: Commit to get hunks of.
+
+    Returns:
+        List of split commit hunks stored as strings where each string
+        represents changes of a single file.
+    """
+    patch = repo.git.show(commit, format="", color="never")
+    hunk_start_re = re.compile(r"diff --git a/.+ b/.+")
+    section_start = 0
+    result = []
+    patch_lines = patch.splitlines()
+    for i, line in enumerate(patch_lines):
+        if hunk_start_re.match(line):
+            section = patch_lines[section_start:i]
+            if section:
+                result.append("\n".join(section))
+            section_start = i
+    # The last section
+    section = patch_lines[section_start:]
+    if section:
+        result.append("\n".join(section))
+    return result
